Updated off of kilobolt framework
Used kilobolt as a model for main game loop, also imported some classes
that were needed and added code to original framework.  Created
player_char, falling_enemy, and GameScreen classes as well as marked
MainActivity as obsolete.  App will not load, not sure as to cause but
guess is memory problems or objects not being drawn.  Will continue to
use log function to debug and get a testable prototype asap.